Update means of controlling broswer
Removed Sikuli library and incorporated Selenium webdriver to perform actions on a web browser. Reason for replacing Sikuli is because it isn't capable of recognizing certain site elements that appear to be images. The incorporation of Selenium allows us to click on actual site elements using xpaths, which is better in the case that a site's image is replaced.