Add seperate class for controlling a website
Created a seperate class, WebClass, for controlling the automated actions that need to be performed on a website. The Order class uses a WebClass object to go into the TacoBell site and place an order. Seperating the two classes improves the readability and reduces redundant calls. WebClass was designed with the ability to be used on different sites.

Changes:
 - WebClass created to contain methods that perform actions on a site like: opening browser into site, clicking on elements, and etc

 - Order class used primarily to call on WebClass object methods to get to a section of a site